+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * `aoc`: New option `package_folder` to create a folder per package when receiving packages.
   * Breaking Changes:
     * `aoc`: `package_folder` defaults to `id`, so a new level of folder is added when receiving a package. Set to `@none:` to revert to previous behavior.
+    * `config`: In `ascp info`: `openssldir` &rarr; `ascp_openssl_dir`, `openssl_version` &rarr; `ascp_openssl_version`, `sdk_ascp_version` &rarr; `ascp_version`
 
 * 4.22.0 2025-06-23